Factor out insert form query from server action
